[TSG] Backporting for 2.1 (pr64) (2.1.17)

+++ app/code/Magento/Backend/Controller/Adminhtml/System/Design/Delete.php
@@ -10,20 +10,25 @@ class Delete extends \Magento\Backend\Controller\Adminhtml\System\Design
 {
     /**
      * @return \Magento\Backend\Model\View\Result\Redirect
+     * @throws \Magento\Framework\Exception\NotFoundException
      */
     public function execute()
     {
+        if (!$this->getRequest()->isPost()) {
+            throw new \Magento\Framework\Exception\NotFoundException(__('Page not found.'));
+        }
+
         $id = $this->getRequest()->getParam('id');
         if ($id) {
-            $design = $this->_objectManager->create('Magento\Framework\App\DesignInterface')->load($id);
+            $design = $this->_objectManager->create(\Magento\Framework\App\DesignInterface::class)->load($id);
 
             try {
                 $design->delete();
-                $this->messageManager->addSuccess(__('You deleted the design change.'));
+                $this->messageManager->addSuccessMessage(__('You deleted the design change.'));
             } catch (\Magento\Framework\Exception\LocalizedException $e) {
-                $this->messageManager->addError($e->getMessage());
+                $this->messageManager->addErrorMessage($e->getMessage());
             } catch (\Exception $e) {
-                $this->messageManager->addException($e, __("You can't delete the design change."));
+                $this->messageManager->addExceptionMessage($e, __("You can't delete the design change."));
             }
         }
         /** @var \Magento\Backend\Model\View\Result\Redirect $resultRedirect */
